Moving to gradle project and changing  package name
