+++ .gitignore
@@ -1,3 +1,6 @@
 .idea/
 *.iml
 target/
+.gradle
+build
+out/
