+++ settings.gradle
@@ -1,0 +1,5 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+rootProject.name = 'camunda-bpm'
